+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.31 branch
 ### 2.31.8
-* added `getid()` and `isGroup()` methods to the `AuthenticationComponent`, with tests.
+* added `getid()` and `isGroup()` methods to the `AuthenticationComponent`, with tests;
+* fixed little bug for `IdentityHelper::isGroup()` method and improved tests.
 
 ### 2.31.7
 * new passwords also require the presence of a capital letter (in addition to the previous rules);